+++ pages/_error.js
@@ -1,6 +1,6 @@
 import { Component } from 'react';
 
-import App from '../components/App';
+import WebsiteApp from '../components/WebsiteApp';
 import initMaze from '../lib/dxmaze';
 // import './_error.css';
 
@@ -11,13 +11,13 @@ class Four04 extends Component {
 
   render() {
     return (
-      <App>
+      <WebsiteApp>
         <div className="error-page container container--lg" id="dxmaze-holder">
           <canvas id="dxmaze" />
           <h1>Some experiments fail.</h1>
           <h2>Page not found!</h2>
         </div>
-      </App>
+      </WebsiteApp>
     );
   }
 }