+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Recreate `Button` component to be more generic
 - Add `MenuIconButton`
 - Add `eslint-plugin-css-modules`
+- Refactor `App` to be more generic and make `WebsiteApp` more customised
 
 ## [1.8.1] - 2020-04-01