+++ CHANGELOG.md
@@ -5,8 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.8.2] - 2020-04-XX
+## [1.8.2] - 2020-04-30
 
+- Add `diary-files` page of new DX Lab experiment
+- Add `DiaryFilesApp` and associated components
 - Add `DXLabLogo`, `SLNSWLogo` components
 - Refactor `Button` to `CTAButton`
 - Recreate `Button` component to be more generic
@@ -21,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `HeaderNav` component
 - Add `Overlay` component
 - Add `Modal` component
+- Add `TextTruncate` component
 
 ## [1.8.1] - 2020-04-01